Create README.md
This is for mobile application that wishes to connect to web based system. This mobile code shared the same database with web base system where you can extract the required data and information needed with simple steps and by using JSON. You must configure so that both platform has the same ip address which can be used to build connection. Every update made in Web Based will be properly displayed on the mobile . An example :  The event set in database will be display on the mobile application